fixed cosmic crystal destroying non-arena blocks
https://github.com/TelepathicGrunt/Bumblezone/issues/431

+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/entities/living/CosmicCrystalEntity.java
@@ -847,7 +847,7 @@ public class CosmicCrystalEntity extends LivingEntity {
 
             if (hitResult instanceof BlockHitResult blockHitResult) {
                BlockState state = this.level().getBlockState(blockHitResult.getBlockPos());
-               if (state.getBlock().getExplosionResistance() < 1500 && !state.is(BlockTags.WITHER_IMMUNE)) {
+               if (state.getBlock().getExplosionResistance() < 1500 && !state.is(BlockTags.WITHER_IMMUNE) && this.getRemovalReason() == null) {
                    this.level().destroyBlock(blockHitResult.getBlockPos(), true);
                }
             }
@@ -1280,7 +1280,9 @@ public class CosmicCrystalEntity extends LivingEntity {
                             blockState.getBlock().getExplosionResistance() < 1500 &&
                             !blockState.is(BlockTags.WITHER_IMMUNE))
                         {
-                            this.level().destroyBlock(mutableBlockPos, true);
+                            if (this.getRemovalReason() == null) {
+                                this.level().destroyBlock(mutableBlockPos, true);
+                            }
                         }
                         else {
                             try {
@@ -1320,7 +1322,9 @@ public class CosmicCrystalEntity extends LivingEntity {
                             blockState.getBlock().getExplosionResistance() < 1500 &&
                             !blockState.is(BlockTags.WITHER_IMMUNE))
                         {
-                            this.level().destroyBlock(mutableBlockPos, true);
+                            if (this.getRemovalReason() == null) {
+                                this.level().destroyBlock(mutableBlockPos, true);
+                            }
                         }
                     }
                 }
